+++ src/core/types/index.ts
@@ -28,4 +28,4 @@ export {
   Language,
   SourceUpAxis,
 } from './dbTypes'
-export type { Dataset, LayerGeometryType } from './datasetTypes'
+export type { Dataset, DatasetViewport, LayerGeometryType } from './datasetTypes'